+++ changelog.md
@@ -1,5 +1,52 @@
 # Changelog
 
+## Version 0.1.3
+
+### Retail
+
+* Reworked BigWigs Profile.
+* Reworked Plater Profile.
+
+### Classic
+
+* Added toc support for both SoM and BCC.
+* Enabled Experience bar since classic does not have an honor bar.
+* Added OmniCD Profile to BCC (no OmniCD version for classic era).
+* Added Plater Profile to classic.
+* Reworked Plater Profile.
+
+## Version 0.1.2
+
+### Retail
+
+* Huge Layout updates.
+* Windtools profile complete.
+* Reworked OmniCD Profile.
+* Reworked WarpDeplete Profile.
+* Added Capping Profile for casual PvP.
+
+### Classic
+
+* Huge Layout updates.
+* Added Capping Profile for casual PvP.
+
+## Version 0.1.1
+
+### Retail
+
+* Lots of changes to the ElvUI Layout.
+* Added Retail check for WindTools settings.
+
+### Classic
+
+* Lots of changes to the ElvUI Layout.
+
 ## Version 0.1
 
+### Retail
+
 * First private beta with Details, ElvUI, OmniCD, Plater, WarpDeplete profiles and some WeakAuras.
+
+### Classic
+
+* First private beta with ElvUI profile.